# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Secure Storage API
  description: API for the self-hosted secure secret sharing platform.
  version: 1.0.0
servers:
  - url: http://localhost:5001
    description: Local Backend Server
paths:
  /api/auth/login:
    get:
      summary: Initiate Google OAuth 2.0 Flow
      description: Redirects the user to the Google login page.
      parameters:
        - name: inviteCode
          in: query
          description: Optional invite code for registration.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '302':
          description: Redirects to the Google OAuth consent page on success, or to the frontend login page with an error code if the invite code is invalid.
          headers:
            Location:
              schema:
                type: string
              description: Google OAuth challenge URL, or frontend URL (e.g. `{frontendUrl}/#/login?error=invite_invalid`).

  /api/auth/callback:
    get:
      summary: OAuth Callback Endpoint
      description: Handles the callback from Google, authenticates the user, registers them if they have a valid matching invite, and redirects back to the frontend.
      responses:
        '302':
          description: Redirects to the configured Frontend URL.
          headers:
            Location:
              schema:
                type: string
              description: >
                Redirect URL on the frontend. Possible values:
                - `{frontendUrl}/` (Successful authentication/registration)
                - `{frontendUrl}/#/login?error=auth_failed` (Google authentication failed)
                - `{frontendUrl}/#/login?error=email_missing` (Google account has no email claim)
                - `{frontendUrl}/#/auth-error?reason=no_account` (User does not exist and no invite code was supplied)
                - `{frontendUrl}/#/register?error=email_mismatch` (Email doesn't match the invite)
                - `{frontendUrl}/#/register?error=already_registered` (Invite email is already registered)
                - `{frontendUrl}/#/register?error=invite_invalid` (Invite is invalid or already used)

  /api/auth/logout:
    get:
      summary: Log Out Current User
      description: Signs the user out of the cookie authentication scheme.
      responses:
        '200':
          description: Successfully logged out.

  /api/auth/me:
    get:
      summary: Get Current User Session Info
      description: Returns the authentication status, email, and ID of the currently logged-in user.
      responses:
        '200':
          description: Session info returned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  isAuthenticated:
                    type: boolean
                  email:
                    type: string
                    nullable: true
                  id:
                    type: string
                    format: uuid
                    nullable: true

  /api/invites:
    post:
      summary: Create a New Invite
      description: Generates a new invite code for a specific target email. Only accessible to authenticated users.
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: The email address of the invitee.
      responses:
        '200':
          description: Invite created successfully. Returns the unique invite code.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    format: uuid
                    description: The generated invite code.
        '400':
          description: Invalid request. Returned when email is missing, target user is already registered, or an active invite already exists for this email.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Specific validation error message.
        '401':
          description: Unauthorized. Current user session is invalid or missing.

    get:
      summary: Get User's Issued Invites
      description: Retrieves a cursor-paginated list of invites issued by the currently authenticated user.
      security:
        - CookieAuth: []
      parameters:
        - name: lastInviteId
          in: query
          description: Cursor ID (Guid v7) of the last invite from the previous page for pagination.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of invites issued by the user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InviteDto'
        '401':
          description: Unauthorized. Current user session is invalid or missing.

  /api/secrets:
    post:
      summary: Create a New Secret
      description: Uploads and encrypts a secret (text or file payload).
      security:
        - CookieAuth: []
      parameters:
        - name: X-Secret-Comment
          in: header
          description: Description or comment for the secret. Must be percent-encoded (URL-encoded) to support non-ASCII characters.
          required: true
          schema:
            type: string
        - name: X-Secret-IsOneTime
          in: header
          description: Set to true if the secret should self-destruct after the first read.
          required: false
          schema:
            type: boolean
            default: false
        - name: X-Secret-ContentType
          in: header
          description: The standard MIME content type of the secret payload (e.g., text/plain, application/pdf, image/png). Defaults to application/octet-stream if missing.
          required: false
          schema:
            type: string
            default: application/octet-stream
        - name: X-Secret-FileName
          in: header
          description: The optional original file name if a file is being uploaded. Must be percent-encoded (URL-encoded) to support non-ASCII characters.
          required: false
          schema:
            type: string
        - name: X-Secret-IV
          in: header
          description: Base64-encoded Initialization Vector (IV) for the encryption.
          required: true
          schema:
            type: string
      requestBody:
        description: Raw encrypted binary stream.
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Secret created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '400':
          description: Missing IV, comment, or invalid content type.
        '401':
          description: Unauthorized.

  /api/secrets/{secretId}:
    get:
      summary: Retrieve a Secret
      description: Returns the encrypted payload and metadata of a secret by its ID. One-time secrets will be marked as burned immediately.
      parameters:
        - name: secretId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Secret retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretDto'
        '404':
          description: Secret not found, expired, or already burned.
    delete:
      summary: Burn (Invalidate) a Secret
      description: Marks a secret as burned (invalidated) immediately, preventing any further retrieval. Only the owner of the secret can perform this action.
      security:
        - CookieAuth: []
      parameters:
        - name: secretId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Secret successfully burned.
        '401':
          description: Unauthorized.

  /api/secrets/my:
    get:
      summary: Get User's Secrets
      description: Retrieves a paginated list of secrets created by the currently authenticated user.
      security:
        - CookieAuth: []
      parameters:
        - name: lastSecretId
          in: query
          description: Cursor ID (Guid v7) of the last secret from the previous page for pagination.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of secrets.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecretSummaryDto'
        '401':
          description: Unauthorized.

components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: .AspNetCore.Cookies
  schemas:
    SecretDto:
      type: object
      properties:
        encryptedData:
          type: string
          format: byte
          description: Base64-encoded encrypted secret content.
        iv:
          type: string
          format: byte
          description: Base64-encoded Initialization Vector.
        contentType:
          type: string
          description: Content type of the secret payload.
        fileName:
          type: string
          nullable: true
          description: Original file name.
        isOneTime:
          type: boolean
          description: Whether this is a burn-on-read secret.
    SecretSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        comment:
          type: string
          nullable: true
        isOneTime:
          type: boolean
        createdAt:
          type: string
          format: date-time
    InviteDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the invite.
        email:
          type: string
          description: Email address of the invitee.
        isUsed:
          type: boolean
          description: Whether the invite has been used.
        issuedByUserId:
          type: string
          format: uuid
          description: The ID of the user who issued the invite.
        usedAt:
          type: string
          format: date-time
          nullable: true
          description: The date and time when the invite was used.
        createdAt:
          type: string
          format: date-time
          description: The date and time when the invite was created.